refactor: 룸 조인 로직을 (tabs) 그룹으로 이동 및 리팩토링
